Later night nov 12
made a lot of headway since last commit. Most of previous complaints are dealt with. 
persisting is the width length not properly taking decimals after toying with most of the db settings... this should be easy! its databasing! all this thing does is take numbers and strings! I am moving on to picture upload tasks now. 

-must figure out how to make list items sortable.