NOV 21 ___This is the master tape
Nearing the end of this day I have fixed a lot of messes and feel I should start a new branch to segment of this phase of the project. There is still a lot to do, but what is here is stable and functional. This amount represents the skeleton of the project with the extra test files removed.